Use a dedicated chip-types.xml file into src/app/zap-templates/zcl (#7873)
* Update chip-types.xml

* Update gen/ folders

+++ src/app/util/af.h
@@ -419,29 +419,29 @@ extern EmberAfDefinedEndpoint emAfEndpoints[];
 /**
  * @brief Macro that takes index of endpoint, and returns Zigbee endpoint
  */
-chip::EndpointId emberAfEndpointFromIndex(uint8_t index);
+chip::EndpointId emberAfEndpointFromIndex(uint16_t index);
 
 /**
  * Returns the index of a given endpoint
  */
-uint8_t emberAfIndexFromEndpoint(chip::EndpointId endpoint);
+uint16_t emberAfIndexFromEndpoint(chip::EndpointId endpoint);
 
 /**
  * Returns the index of a given endpoint; Does not ignore disabled endpoints
  */
-uint8_t emberAfIndexFromEndpointIncludingDisabledEndpoints(chip::EndpointId endpoint);
+uint16_t emberAfIndexFromEndpointIncludingDisabledEndpoints(chip::EndpointId endpoint);
 
 /**
  * Returns the endpoint index within a given cluster (Client-side),
  * looking only for standard clusters.
  */
-uint8_t emberAfFindClusterClientEndpointIndex(chip::EndpointId endpoint, chip::ClusterId clusterId);
+uint16_t emberAfFindClusterClientEndpointIndex(chip::EndpointId endpoint, chip::ClusterId clusterId);
 
 /**
  * Returns the endpoint index within a given cluster (Server-side),
  * looking only for standard clusters.
  */
-uint8_t emberAfFindClusterServerEndpointIndex(chip::EndpointId endpoint, chip::ClusterId clusterId);
+uint16_t emberAfFindClusterServerEndpointIndex(chip::EndpointId endpoint, chip::ClusterId clusterId);
 
 /**
  * @brief Macro that takes index of endpoint, and returns device Id for it
@@ -466,12 +466,12 @@ uint8_t emberAfFindClusterServerEndpointIndex(chip::EndpointId endpoint, chip::C
 /**
  * @brief Returns the total number of endpoints (dynamic and pre-compiled).
  */
-uint8_t emberAfEndpointCount(void);
+uint16_t emberAfEndpointCount(void);
 
 /**
  * @brief Returns the number of pre-compiled endpoints.
  */
-uint8_t emberAfFixedEndpointCount(void);
+uint16_t emberAfFixedEndpointCount(void);
 
 /**
  * Data types are either analog or discrete. This makes a difference for
@@ -667,7 +667,7 @@ bool emberAfEndpointEnableDisable(chip::EndpointId endpoint, bool enable);
 /**
  * @brief Determine if an endpoint at the specified index is enabled or disabled
  */
-bool emberAfEndpointIndexIsEnabled(uint8_t index);
+bool emberAfEndpointIndexIsEnabled(uint16_t index);
 
 /**
  * @brief Returns true if a given ZCL data type is a string type.
